fixed accept invite
