+++ screens/team-details/team-details.js
